celbpf: remove arg indirection
Prior to this change, the user was required to declare the args that
they referenced within their CelExpr operator matchArgs selector. Within
the "args" section of the matchArgs selector, they would list indexes
that signified the offset within the "args" section of the tracing
policy.

This is inconvenient and redundant, as the referenced args can be
determined directly by parsing the CEL expressing itself.

This also created an unneeded layer of indirection. Ultimately, we need
a way to understand which arg within the "args" section of the hooked
function is being referenced. This change makes it so the numeric suffix
of the args in the CEL expression signify the offset within the args
configuration of the hooked function.

Previously, the numeric suffix of the args referenced in the CEL
expression were used as an offset into the "args" section of the
matchArgs selector. The index located at this offset, was then used as
the offset into the "args" section of the hook point.

+++ pkg/celbpf/codegen.go
@@ -140,8 +140,8 @@ func (g *codeGenerator) emitU32(reg asm.Register, regTy *cgTypes.Type) error {
 	return nil
 }
 
-func (g *codeGenerator) pushArg(argTy *cgTypes.Type, argOffset int, tmp1, tmp2 asm.Register) error {
-	off := argOffset * 8
+func (g *codeGenerator) pushArg(argTy *cgTypes.Type, argOffset uint16, tmp1, tmp2 asm.Register) error {
+	off := int(argOffset * 8)
 	if off > math.MaxInt16 {
 		return fmt.Errorf("offset %d overflows 16-bits", off)
 	}
